+++ ee/packages/ui-theming/src/PaletteStyleTag.tsx
@@ -7,15 +7,15 @@ import { filterOnlyChangedColors } from './helpers/filterOnlyChangedColors';
 import { defaultPalette } from './palette';
 import { darkPalette } from './paletteDark';
 import { useThemeMode } from './hooks/useThemeMode';
-import { createStyleContainer } from './helpers/createStyleContainer';
+import { useCreateStyleContainer } from './hooks/useCreateStyleContainer';
 
 export const PaletteStyleTag = memo((): ReactElement | null => {
 	const [, , theme] = useThemeMode();
 
-	if (theme !== 'dark') {
-		return null;
-	}
-	const palette = convertToCss(filterOnlyChangedColors(defaultPalette, darkPalette), '.rcx-content--main');
+	const palette =
+		theme === 'dark'
+			? convertToCss(filterOnlyChangedColors(defaultPalette, darkPalette), '.rcx-content--main')
+			: convertToCss(filterOnlyChangedColors(defaultPalette, {}), '.rcx-content--main');
 
-	return createPortal(palette, createStyleContainer('main-palette'));
+	return createPortal(palette, useCreateStyleContainer('main-palette'));
 });
