+++ ee/packages/ui-theming/src/SidebarPaletteStyleTag.tsx
@@ -5,7 +5,7 @@ import { createPortal } from 'react-dom';
 import { defaultSidebarPalette } from './sidebarPalette';
 import { darkPalette } from './paletteDark';
 import { convertToCss } from './helpers/convertToCss';
-import { createStyleContainer } from './helpers/createStyleContainer';
+import { useCreateStyleContainer } from './hooks/useCreateStyleContainer';
 
 export const SidebarPaletteStyleTag = memo((): ReactElement | null => {
 	// Commented code below: sidebar palette currently the same in both themes.
@@ -18,5 +18,5 @@ export const SidebarPaletteStyleTag = memo((): ReactElement | null => {
 
 	const palette = convertToCss({ ...darkPalette, ...defaultSidebarPalette }, '.rcx-sidebar--main');
 
-	return createPortal(palette, createStyleContainer('sidebar-palette'));
+	return createPortal(palette, useCreateStyleContainer('sidebar-palette'));
 });
