--- ee/packages/ui-theming/src/helpers/createStyleContainer.ts
@@ -1,7 +1,0 @@
-export const createStyleContainer = (id: string) => {
-	const refElement = document.getElementById('css-theme') || document.head.lastChild;
-	const styleElement = document.createElement('style');
-	styleElement.setAttribute('id', id);
-	document.head.insertBefore(styleElement, refElement);
-	return document.getElementById(id) || document.head.appendChild(document.createElement('style'));
-};
